+++ Runtime/InternalBridge/FocusControllerExtensionsBridge.cs
@@ -32,7 +32,7 @@ namespace Unity.AppUI.Bridge
             return k_FocusNextInDirection.Invoke(controller, new object[] { currentFocusable, direction }) as Focusable;
         }
 #else
-        internal static Focusable FocusNextInDirection(this FocusController controller, FocusChangeDirection direction)
+        internal static Focusable FocusNextInDirection(this FocusController controller, Focusable currentFocusable, FocusChangeDirection direction)
         {
             return k_FocusNextInDirection.Invoke(controller, new object[] { direction }) as Focusable;
         }