+++ CHANGELOG.md
@@ -4,6 +4,27 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0-pre.17] - 2025-01-29
+
+### Changed
+
+- Replaced ClickEvent action in MenuItem builder by an EventBase action
+
+### Fixed
+
+- Fixed styling of MenuItem element
+- Fixed some SplitView styling issues when nesting SplitViews in a visual tree.
+- Fixed context propoagation per key
+
+### Added
+
+- Added scoped service provider implementation.
+- Added `showExpandButtons` property to the `SplitView` UI element.
+
+### Removed
+
+- Removed EventBaseExtensionsBridge class
+
 ## [2.0.0-pre.16] - 2025-01-15
 
 ### Removed